Move Unit for better encaptulation.
